(maint) refactor DscProviderHelpers
 - Given there is no more MOF provider, it's not
   necessary any longer to have a separate
   helpers class shared amongst multiple providers

   Simplify the implementation by moving the few
   small methods into the powershell provider
   rather than mixing them in.
