+++ lib/puppet/provider/base_dsc/powershell.rb
@@ -1,14 +1,37 @@
-require 'puppet_x/msutter/helpers/dsc_provider_helpers'
-
 Puppet::Type.type(:base_dsc).provide(:powershell) do
-
-  include PuppetX::Msutter::Helpers::DscProviderHelpers
-
   confine :operatingsystem => :windows
   defaultfor :operatingsystem => :windows
+
+  commands :powershell =>
+    if File.exists?("#{ENV['SYSTEMROOT']}\\sysnative\\WindowsPowershell\\v1.0\\powershell.exe")
+      "#{ENV['SYSTEMROOT']}\\sysnative\\WindowsPowershell\\v1.0\\powershell.exe"
+    elsif File.exists?("#{ENV['SYSTEMROOT']}\\system32\\WindowsPowershell\\v1.0\\powershell.exe")
+      "#{ENV['SYSTEMROOT']}\\system32\\WindowsPowershell\\v1.0\\powershell.exe"
+    else
+      'powershell.exe'
+    end
+
   desc <<-EOT
 Applies DSC Resources by generating a configuration file and applying it.
 EOT
+
+  def dsc_parameters
+    resource.parameters_with_value.select do |p|
+      p.name.to_s =~ /dsc_/
+    end
+  end
+
+  def template_path
+    File.expand_path('../../../../puppet_x/msutter/templates', __FILE__)
+  end
+
+  def vendored_modules_path
+    File.expand_path(File.join(__FILE__, '..', '..', '..', '..', '..', 'vendor','DSCResources')).gsub(/\//,'\\')
+  end
+
+  def powershell_args
+    ['-NoProfile', '-NonInteractive', '-NoLogo', '-ExecutionPolicy', 'Bypass', '-Command']
+  end
 
   def exists?
     script_content = ps_script_content('test')
